requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
Django==5.1.2
djangorestframework==3.15.2
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.0
python-decouple==3.8
sqlparse==0.5.1
typing_extensions==4.12.2
vt-py==0.18.4
yarl==1.17.1
